+++ docs/CHANGELOG.rst
@@ -14,6 +14,7 @@ Changed
 -------
 - **BACKWARD INCOMPATIBLE:** Require ``Django 5.2.x`` and bump version of
   dependencies
+- Move full-text search to its own filter
 
 
 ===================